requirements.txt
numpy
pillow
matplotlib
trimesh
pyrender
open3d
rembg
torch

